% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_symp_cat_weights.R
\name{calculate_symp_cat_weights}
\alias{calculate_symp_cat_weights}
\title{Calculate symptom category weights}
\usage{
calculate_symp_cat_weights(joint_dataset)
}
\arguments{
\item{joint_dataset}{a data frame that includes symp_type_cat}
}
\value{
A data frame with one row for each symp_type_cat and the
corresponding weight.
}
\description{
Get weights for each symptom category based on observed proportion
}
