% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_stan_variables.R
\name{create_stan_variables}
\alias{create_stan_variables}
\title{Create Stan model indicator variables}
\usage{
create_stan_variables(data, parameters_config)
}
\arguments{
\item{data}{output of stan_preprocess}

\item{parameters_config}{parameters_config.yml}
}
\value{
list of dataframes with additional variables needed for stan
}
\description{
for easy running of the stan model, indicator variables
and such data manipulations are done outside the model
}
