% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_symptom_improvement.R
\name{estimate_symptom_improvement}
\alias{estimate_symptom_improvement}
\title{Simulate symptom improvement time}
\usage{
estimate_symptom_improvement(si_scale, si_shape, max_si)
}
\arguments{
\item{si_scale}{escale parameter for weibull}

\item{si_shape}{shape parameter for weibull}

\item{max_si}{maximum time to symptom improvement}
}
\value{
estimated symptom improvement
}
\description{
Simulates symptom improvment using a discreteWeibull
}
