% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_posteriors_to_df.R
\name{extract_posteriors_to_df}
\alias{extract_posteriors_to_df}
\title{Extract Stan output posteriors to data frames}
\usage{
extract_posteriors_to_df(
  vl_stan_fit,
  parameters_config,
  extract_symp_improv_params = TRUE
)
}
\arguments{
\item{vl_stan_fit}{viral load stan fit}

\item{parameters_config}{parameters config object}

\item{extract_symp_improv_params}{whether to extract
symptom parameters from stan fit object
(i.e., weibull proportional hazards coefs)}
}
\value{
dataframe of individual-level parameters +
dataframe of pooled parameters
}
\description{
extract the pooled posterior samples from the stan fit
and output to a dataframe
}
