% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_stan_input_list.R
\name{generate_stan_input_list}
\alias{generate_stan_input_list}
\title{Make input list for Stan}
\usage{
generate_stan_input_list(data, parameters_config)
}
\arguments{
\item{data}{output of stan_preprocess}

\item{parameters_config}{parameters_config.yml}
}
\value{
data list ready for stan input
}
\description{
stan expects a list of variables -- turn a dataframe
that is ready for stan into a list which also has all the
parameters config information
}
