% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_antigen_prob_detection.R
\name{get_antigen_prob_detection}
\alias{get_antigen_prob_detection}
\title{Calculate probability of detection by antigen testing}
\usage{
get_antigen_prob_detection(
  test_t,
  t,
  antigen_sens,
  symp_improv,
  symp_improv_adjust,
  iso_lag_adjust
)
}
\arguments{
\item{test_t}{Times at which testing is indicated}

\item{t}{Vector of times at which antigen sensitivity is known}

\item{antigen_sens}{Vector of antigen test sensitivities at time t}

\item{symp_improv}{Observed time of symptom improvement}

\item{symp_improv_adjust}{Presumed time between actual and observed symptom
improvement time}

\item{iso_lag_adjust}{Presumed time between actual symptom onset and end of
"Day 0", which is t=0}
}
\value{
probability of antigen detection
}
\description{
The probability that a case will be detected through antigen
testing is calculated as 1 minus the product of the probabilities
that each test will be negative. This is one probability per case;
it is not time-varying. A test is not taken if the patient's symptoms
have already improved.
}
