% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dt_mid_approx.R
\name{get_dt_mid_approx}
\alias{get_dt_mid_approx}
\title{Calculate dt accounting for mid-point approximation}
\usage{
get_dt_mid_approx(t)
}
\arguments{
\item{t}{A vector of time points of viral load observations.}
}
\value{
A vector of dt values the same length as t,
except with start and end values set equal to 1/2 * dt.
}
\description{
Make a vector of the length of time intervals (i.e., dt), but with the
first and last values halved.
}
\examples{
# Here dt is 0.1; this should return vector of length 31
# where values are 0.1, except first and last values are 0.05
dt_mid_approx_fx(seq(from = -1, to = 2, by = 0.1))
}
