% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_infectiousness_averted.R
\name{get_infectiousness_averted}
\alias{get_infectiousness_averted}
\title{Percentage of infectiousness averted}
\usage{
get_infectiousness_averted(I_guid, I_null)
}
\arguments{
\item{I_guid}{infectiousness timeseries when following guidance}

\item{I_null}{infectiousness timeseries when no intervention}
}
\value{
value of the percentage of infectiousness averted
}
\description{
Calculate the percentage reduction in onward transmission potential when
following guidance versus in the absence of any intervention
}
