% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_antigen_positive.R
\name{prob_antigen_positive}
\alias{prob_antigen_positive}
\title{Calculate probability an antigen test is positive}
\usage{
prob_antigen_positive(logVL, antigen_50, sigma_antigen)
}
\arguments{
\item{logVL}{underlying log viral load}

\item{antigen_50}{logVL at which antigen positive probability is 50\%}

\item{sigma_antigen}{"steepness"/variability in sigmoid curve
as antigen probability increases with increasing logVL}
}
\value{
probability antigen positive for each given logVL
}
\description{
Probability an antigen test is positive depends on log viral load and
parameters that relate viral load to test sensitivity
}
