% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_culture_positive.R
\name{prob_culture_positive}
\alias{prob_culture_positive}
\title{Calculate probability a culture assay is positive}
\usage{
prob_culture_positive(logVL, t, culture_50, sigma_culture, culture_beta, tp)
}
\arguments{
\item{logVL}{underlying log viral load}

\item{t}{time at which logVL measurements were taken}

\item{culture_50}{logVL at which culture positive
probability is 50\% at t = tp}

\item{sigma_culture}{"steepness"/variability in sigmoid curve
as culture probability increases with increasing logVL}

\item{culture_beta}{change in log-odds of culture positive as time
since peak VL increases}

\item{tp}{peak viral load time}
}
\value{
probability of testing culture positive based on VL and time
relative to peak viral load
}
\description{
Probability a culture assay is positive depends on log viral load and time
relative to peak viral load, plus parameters that relate these inputs to
sensitivity.
}
