% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_ids_overall_symp_cat.R
\name{sample_ids_overall_symp_cat}
\alias{sample_ids_overall_symp_cat}
\title{Weighted sample to reflect frequency of symptom categories}
\usage{
sample_ids_overall_symp_cat(symp_id_df, symp_cat_wts, n_overall_sample)
}
\arguments{
\item{symp_id_df}{a data frame relating posterior IDs and symptom categories}

\item{symp_cat_wts}{a data frame with weights for each symptom category}

\item{n_overall_sample}{number of samples in the overall "category"}
}
\value{
a dataframe of posterior IDs that have been sampled
}
\description{
Weighted sample to reflect frequency of symptom categories
}
