% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_intrinsic_infectiousness.R
\name{simulate_intrinsic_infectiousness}
\alias{simulate_intrinsic_infectiousness}
\title{Simulate intrinsic time-varying infectiousness}
\usage{
simulate_intrinsic_infectiousness(
  t,
  vl_curve,
  prob_antigen_positive,
  prob_culture_positive,
  infec_threshold,
  transform_fun
)
}
\arguments{
\item{t}{time vector}

\item{vl_curve}{time-varying log viral load values for one person}

\item{prob_antigen_positive}{probabilities of antigen assay testing positive}

\item{prob_culture_positive}{probabilities of having culturable virus}

\item{infec_threshold}{logVL to use as base of infectiousness triangle}

\item{transform_fun}{viral load to infectiousness transformation (logVL,
antigen, culture, logVL_x_antigen, or logVL_x_culture)}
}
\value{
a time series of infectiousness, I(t), multiplied by time step, dt;
summing these values is numerical integration of area under infectiousness
curve
}
\description{
Create a time series of an individual's intrinsic infectiousness (i.e.,
their potential for infecting others in the absence of interventions).
Multiple functional forms are supported to relate log viral load, antigen or
culture positivity, and combinations thereof to infectiousness.
}
