% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_isolation_current.R
\name{simulate_isolation_current}
\alias{simulate_isolation_current}
\title{Simulate previous guidance for isolation and post-isolation precautions}
\usage{
simulate_isolation_current(
  It,
  t,
  prob_antigen_positive,
  symp_type_cat,
  symp_improv,
  iso_lag,
  test_count,
  test_interval,
  min_iso_days_moderate,
  min_iso_days_mild,
  min_total_precaution_days,
  iso_eff,
  post_iso_eff,
  symp_improv_adjust,
  iso_lag_adjust
)
}
\arguments{
\item{It}{Vector of time-varying infectiousness obtained from
\code{simulate_intrinsic_infectiousness()}}

\item{t}{Vector of time values}

\item{prob_antigen_positive}{Vector of probabilities of being antigen
positive}

\item{symp_type_cat}{An integer indicating person's symptom category}

\item{symp_improv}{Observed time of symptom improvement}

\item{iso_lag}{Lag from symptom onset to isolation}

\item{test_count}{Number of antigen tests taken (if remaining symptomatic)}

\item{test_interval}{Time between antigen tests}

\item{min_iso_days_moderate}{Minimum days in isolation from symptom onset
for persons with moderate symptoms}

\item{min_iso_days_mild}{Minimum days in isolation from symptom onset
for persons with mild symptoms}

\item{min_total_precaution_days}{Minimum total days spent in isolation and
post-isolation precautions combined}

\item{iso_eff}{Efficacy of isolation in reducing transmission potential}

\item{post_iso_eff}{Efficacy of post-isolation precautions in reducing
transmission potential}

\item{symp_improv_adjust}{Presumed time between actual and observed symptom
improvement time}

\item{iso_lag_adjust}{Presumed time between actual symptom onset and end of
"Day 0", which is t=0}
}
\value{
A list consisting of the following:
\itemize{
\item antigen_detect_prob: the probability the individual's infection is
detected by antigen testing
\item p_iso: vector of time-varying probability that a person is in isolation
\item p_post_iso: vector of time-varying probability that a person is taking
post-isolation precautions
\item I_t: time-varying transmission potential (area under the curve) when
following the guidance
\item averted : percentage of transmission potential averted by following the
guidance
}
}
\description{
Simulate how much transmission potential may be averted by following
CDC's guidance for duration of isolation and post-isolation precautions for
persons with COVID-19, based on the guidance that existed prior to
2024-03-01. (The reference to "current" in the function name is because the
initial version of this code was prepared prior to 2024-03-01.)
}
