% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_isolation_proposed.R
\name{simulate_isolation_proposed}
\alias{simulate_isolation_proposed}
\title{Simulate updated guidance for isolation and post-isolation precautions}
\usage{
simulate_isolation_proposed(
  It,
  t,
  symp_improv_in,
  symp_improv_adjust,
  iso_lag,
  iso_lag_adjust,
  min_post_iso_days,
  iso_eff,
  post_iso_eff
)
}
\arguments{
\item{It}{Vector of time-varying infectiousness obtained from
\code{simulate_intrinsic_infectiousness()}}

\item{t}{Vector of time values}

\item{symp_improv_in}{Observed time of symptom improvement}

\item{symp_improv_adjust}{Presumed time between actual and observed symptom
improvement time}

\item{iso_lag}{Lag from symptom onset to isolation}

\item{iso_lag_adjust}{Presumed time between actual symptom onset and end of
"Day 0", which is t=0}

\item{min_post_iso_days}{Minimum days in post-isolation precautions}

\item{iso_eff}{Efficacy of isolation in reducing transmission potential}

\item{post_iso_eff}{Efficacy of post-isolation precautions in reducing
transmission potential}
}
\value{
A list consisting of the following:
\itemize{
\item antigen_detect_prob: the probability the individual's infection is
detected by antigen testing
\item p_iso: vector of time-varying probability that a person is in isolation
\item p_post_iso: vector of time-varying probability that a person is taking
post-isolation precautions
\item I_t: time-varying transmission potential (area under the curve) when
following the guidance
\item averted : percentage of transmission potential averted by following the
guidance
}
}
\description{
Simulate how much transmission potential may be averted by following
CDC's guidance for duration of isolation and post-isolation precautions for
persons with COVID-19, based on the guidance the updated guidance from
2024-03-01. (The reference to "proposed" in the function name is because the
initial version of this code was prepared prior to 2024-03-01.)
}
