% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_preprocess.R
\name{stan_preprocess}
\alias{stan_preprocess}
\title{Preprocess data to input into Stan model}
\usage{
stan_preprocess(
  data_to_prep,
  parameters_config,
  dataset_name,
  at_least_one_test = TRUE
)
}
\arguments{
\item{data_to_prep}{INHERENT/RVTN data to prep for stan}

\item{parameters_config}{a config specifying special values of this dataset}

\item{dataset_name}{data set being analyzed here's name to match config}

\item{at_least_one_test}{only include days for which there is >= 1 test?}
}
\value{
a list that contains 1) processed data and 2) a mermaid diagram
}
\description{
Data processing and associated checks to prepare it for the Stan model
}
