% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangle_vl.R
\name{triangle_vl}
\alias{triangle_vl}
\title{Get viral load values from triangle parameters}
\usage{
triangle_vl(t, dp, tp, wp, wr)
}
\arguments{
\item{t}{times at which viral load should be estimated}

\item{dp}{peak value of viral load -- assumes that peak VL > LOD}

\item{tp}{time of peak vira load}

\item{wp}{proliferation time (time from logVL = 0 up to peak)}

\item{wr}{clearance time (time from peak back down to logVL = 0)}
}
\value{
triangular viral load at each time t
}
\description{
Calulate (log) viral load at given times based on parameters for
"triangular" VL function
}
