% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vl_fits_to_pdf.R
\name{vl_fits_to_pdf}
\alias{vl_fits_to_pdf}
\title{Plot viral load fits}
\usage{
vl_fits_to_pdf(
  vl_stan_fit,
  data_list,
  parameters_config,
  main_name,
  figure_dir,
  data_output_dir
)
}
\arguments{
\item{vl_stan_fit}{fitted stan fit object}

\item{data_list}{underlying data used for \code{vl_stan_fit}}

\item{parameters_config}{config file}

\item{main_name}{name under which output data will be saved}

\item{figure_dir}{directory where PDF fit file will be saved}

\item{data_output_dir}{directory where data output will be saved}
}
\value{
saves PDF plots and data frames of observed data and corresponding
fitted viral load trajectories
}
\description{
Plot individual-level VL fits to a PDF and save underlying data; these plots
are used to check model performance by visual inspection
}
